move random pairs to convenience
