+++ dedupe/core.py
@@ -4,9 +4,7 @@
 import itertools
 import tempfile
 import os
-import random
 import collections
-import warnings
 import functools
 import multiprocessing
 import multiprocessing.dummy
@@ -38,77 +36,6 @@ class BlockingError(Exception):
 
 
 _Queue = Union[multiprocessing.dummy.Queue, multiprocessing.Queue]
-IndicesIterator = Iterator[Tuple[int, int]]
-
-
-def randomPairs(n_records: int, sample_size: int) -> IndicesIterator:
-    """
-    Return random combinations of indices for a square matrix of size n
-    records. For a discussion of how this works see
-    http://stackoverflow.com/a/14839010/98080
-
-    """
-    n: int = n_records * (n_records - 1) // 2
-
-    if not sample_size:
-        return iter([])
-    elif sample_size >= n:
-        random_pairs = numpy.arange(n)
-    else:
-        try:
-            random_pairs = numpy.array(random.sample(range(n), sample_size),
-                                       dtype=numpy.uint)
-        except OverflowError:
-            return randomPairsWithReplacement(n_records, sample_size)
-
-    b: int = 1 - 2 * n_records
-
-    i = (-b - 2 * numpy.sqrt(2 * (n - random_pairs) + 0.25)) // 2
-    i = i.astype(numpy.uint)
-
-    j = random_pairs + i * (b + i + 2) // 2 + 1
-    j = j.astype(numpy.uint)
-
-    return zip(i, j)
-
-
-def randomPairsMatch(n_records_A: int, n_records_B: int, sample_size: int) -> IndicesIterator:
-    """
-    Return random combinations of indices for record list A and B
-    """
-    n: int = n_records_A * n_records_B
-
-    if not sample_size:
-        return iter([])
-    elif sample_size >= n:
-        random_pairs = numpy.arange(n)
-    else:
-        random_pairs = numpy.array(random.sample(range(n), sample_size))
-
-    i, j = numpy.unravel_index(random_pairs, (n_records_A, n_records_B))
-
-    return zip(i, j)
-
-
-def randomPairsWithReplacement(n_records: int, sample_size: int) -> IndicesIterator:
-    # If the population is very large relative to the sample
-    # size than we'll get very few duplicates by chance
-    warnings.warn("The same record pair may appear more than once in the sample")
-
-    try:
-        random_indices = numpy.random.randint(n_records,
-                                              size=sample_size * 2)
-    except (OverflowError, ValueError):
-        max_int: int = numpy.iinfo('int').max
-        warnings.warn("Asked to sample pairs from %d records, will only sample pairs from first %d records" % (n_records, max_int))
-
-        random_indices = numpy.random.randint(max_int,
-                                              size=sample_size * 2)
-
-    random_indices = random_indices.reshape((-1, 2))
-    random_indices.sort(axis=1)
-
-    return ((p.item(), q.item()) for p, q in random_indices)
 
 
 class ScoreDupes(object):
